[tlsconfig] Move tlsconfig (back) to common.
We moved tlsconfig from internal/tlsconfig to common/tlsconfig in
https://github.com/cloudprober/cloudprober/pull/592, but users need
access to tlsconfig protos to programmatically generate Cloudprober
configs.
