+++ config/proto/config.proto
@@ -2,7 +2,7 @@ syntax = "proto2";
 
 package cloudprober;
 
-import "github.com/cloudprober/cloudprober/internal/tlsconfig/proto/config.proto";
+import "github.com/cloudprober/cloudprober/common/tlsconfig/proto/config.proto";
 import "github.com/cloudprober/cloudprober/probes/proto/config.proto";
 import "github.com/cloudprober/cloudprober/internal/rds/server/proto/config.proto";
 import "github.com/cloudprober/cloudprober/internal/servers/proto/config.proto";
