+++ internal/rds/client/proto/config.proto
@@ -10,7 +10,7 @@ syntax = "proto2";
 package cloudprober.rds;
 
 import "github.com/cloudprober/cloudprober/internal/oauth/proto/config.proto";
-import "github.com/cloudprober/cloudprober/internal/tlsconfig/proto/config.proto";
+import "github.com/cloudprober/cloudprober/common/tlsconfig/proto/config.proto";
 import "github.com/cloudprober/cloudprober/internal/rds/proto/rds.proto";
 
 option go_package = "github.com/cloudprober/cloudprober/internal/rds/client/proto";
