+++ internal/rds/kubernetes/proto/config.proto
@@ -25,7 +25,7 @@ syntax = "proto2";
 
 package cloudprober.rds.kubernetes;
 
-import "github.com/cloudprober/cloudprober/internal/tlsconfig/proto/config.proto";
+import "github.com/cloudprober/cloudprober/common/tlsconfig/proto/config.proto";
 
 option go_package = "github.com/cloudprober/cloudprober/internal/rds/kubernetes/proto";
