+++ probes/http/proto/config.proto
@@ -3,7 +3,7 @@ syntax = "proto2";
 package cloudprober.probes.http;
 
 import "github.com/cloudprober/cloudprober/internal/oauth/proto/config.proto";
-import "github.com/cloudprober/cloudprober/internal/tlsconfig/proto/config.proto";
+import "github.com/cloudprober/cloudprober/common/tlsconfig/proto/config.proto";
 import "github.com/cloudprober/cloudprober/metrics/payload/proto/config.proto";
 
 option go_package = "github.com/cloudprober/cloudprober/probes/http/proto";
